Audit with ruff
